+++ src/types/frontend.ts
@@ -6,6 +6,13 @@ export interface CompilerInfo {
 	options ?: any;
 };
 
+export interface CompileOptions {
+	[option: string]: any;
+	impl?: boolean;
+	extOnly?: boolean;
+	"odata-version"?: string | number;
+};
+
 export interface OutputData {
 	filename: string;
 	code: string;